Refactor: Refatored UseCases class

+++ src/useCases/deleteProductUseCase.js
@@ -5,7 +5,27 @@ class DeleteProductUseCase {
 
     async execute(productId) {
 
-        return await this.productRepository.deleteProduct(productId);
+        const product = await this.productRepository.getProductById(productId);
+
+
+        if (product == undefined) {
+
+            const messageNoProducts = {
+                code: 404,
+                message: `Product ID: ${productId} not found in the database`
+            }
+          
+            return messageNoProducts;
+        }
+
+        await this.productRepository.deleteProduct(productId);
+
+        const productDeleteResponse = {
+            code: 200,
+            product: `Product ID: ${productId} deleted successfully`
+        }
+        
+        return productDeleteResponse;
     }
 }
